Fix technical bug
Previously, when evaluating apps, we had `clenv ++ env` as the environment to evaluate the body in. This is actually wrong; when evaluating a function, we should _not_ be able to see local values with the same names as free variables of the closure. However, this didn't actually cause observable misbehavior because of type checking. A term that would be affected by this mistake would be thrown out by the call to `check`. Regardless, this is important to get right in weakly/dynamically typed languages.